+++ src/plugins/calendar_view/icon.png
