+++ src/plugins/calendar_view/settings.html
